2020-02-practioneer/scripts/02/alarm/requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
appnope==0.1.0
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pyrsistent==0.15.7
python-dateutil==2.8.1
pyzmq==18.1.1
qtconsole==4.6.0
Send2Trash==1.5.0
slackclient==2.5.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
virtualenv==20.0.4
virtualenv-clone==0.5.3
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yarl==1.4.2
zipp==3.0.0
